fix cert renewals not using the correct dns-01 config
The background maintenance process is started when certmagic creates a
cache. But because I call NewDefault(), I end up using the default
cache, which uses the default cert config for cert renewals. But I DID
NOT pass the DNS-01 solver to the default cert config :/. Only the
staticCertConfig which I manually instantiated has access to the DNS-01
solver. That's why background cert renewals fail (the cert config does
not have DNS-01), but manually restarting the server fixes it (it makes
an initial call to staticCertConfig.ManageSync, which does have DNS-01).

The solution is to ensure that the staticCertConfig
uses a cache which uses staticCertConfig for renewals. You have to do a
bit of circular instantiation by defining a nil staticCertConfig pointer
first, instantiating a cache that returns the staticCertConfig pointer,
then actually instantiate the staticCertConfig pointer using the cache.

I wouldn't have known any of this without Codex (ChatGPT). I don't know
how the hell any of this works. I only pointed it at the locally-cloned
certmagic code as well as mine and asked it to figure it out. If ChatGPT
didn't exist, I'd never have been able to solve this. I sure hope this
works.